Refactor a bit the validation of Statistics, enable unit test
