+++ src/Sylius/Bundle/ApiBundle/Controller/GetStatisticsAction.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ApiBundle\Controller;
 
-use ApiPlatform\Symfony\Validator\Exception\ValidationException;
 use Sylius\Bundle\ApiBundle\Query\GetStatistics;
 use Sylius\Bundle\ApiBundle\Validator\Constraints;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -25,6 +24,8 @@ use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints as SymfonyConstraints;
+use Symfony\Component\Validator\ConstraintViolationListInterface;
+use Symfony\Component\Validator\Context\ExecutionContextInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 final class GetStatisticsAction
@@ -58,7 +59,7 @@ final class GetStatisticsAction
 
         $violations = $this->validator->validate($parameters, $this->constraints);
         if (count($violations) > 0) {
-            throw new ValidationException($violations);
+            return $this->createValidationErrorResponse($violations);
         }
 
         $interval = $parameters['interval'];
@@ -102,7 +103,9 @@ final class GetStatisticsAction
                     new SymfonyConstraints\DateTime('Y-m-d\TH:i:s', message: 'sylius.date_time.invalid'),
                 ],
             ]),
-            new SymfonyConstraints\Expression(expression: 'value["startDate"] < value["endDate"]', message: 'sylius.statistics.end_date.invalid'),
+            new SymfonyConstraints\Callback(function ($data, ExecutionContextInterface $context) {
+                $this->validateDateRange($data, $context);
+            }),
         ];
     }
 
@@ -119,5 +122,29 @@ final class GetStatisticsAction
         }
 
         return $intervals;
+    }
+
+    private function validateDateRange($data, ExecutionContextInterface $context): void
+    {
+        if ((isset($data['startDate'], $data['endDate'])) && $data['startDate'] >= $data['endDate']) {
+            $context->buildViolation('sylius.statistics.end_date.invalid')
+                ->addViolation();
+        }
+    }
+
+    private function createValidationErrorResponse(ConstraintViolationListInterface $violations): JsonResponse
+    {
+        $errors = [];
+        foreach ($violations as $violation) {
+            $errors[] = [
+                'propertyPath' => $violation->getPropertyPath(),
+                'message' => $violation->getMessage(),
+            ];
+        }
+
+        return new JsonResponse([
+            'status' => Response::HTTP_UNPROCESSABLE_ENTITY,
+            'violations' => $errors,
+        ], Response::HTTP_UNPROCESSABLE_ENTITY);
     }
 }
